feat: before taking a Backup check if all required indices are present
Before taking a backup check in ES/OS which indices are present. If a
required index is missing, return an error.
If a non-required index is missing, just skip it. If there is no index
for a particular part, skip that part altogether.
Skipping empty parts avoids doing a snapshot of the entire database

relates #26618

+++ webapps-backup/src/main/java/io/camunda/webapps/backup/BackupServiceImpl.java
@@ -11,6 +11,7 @@ import io.camunda.webapps.backup.BackupException.*;
 import io.camunda.webapps.backup.repository.BackupRepositoryProps;
 import io.camunda.webapps.schema.descriptors.backup.BackupPriorities;
 import io.camunda.webapps.schema.descriptors.backup.SnapshotIndexCollection;
+import io.camunda.zeebe.util.VisibleForTesting;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Queue;
@@ -32,8 +33,6 @@ public class BackupServiceImpl implements BackupService {
 
   private final BackupRepository repository;
 
-  private final List<SnapshotIndexCollection> indexPatternsOrdered;
-
   public BackupServiceImpl(
       final Executor threadPoolTaskExecutor,
       final BackupPriorities backupPriorities,
@@ -42,7 +41,6 @@ public class BackupServiceImpl implements BackupService {
       final DynamicIndicesProvider dynamicIndicesProvider) {
     this.threadPoolTaskExecutor = threadPoolTaskExecutor;
     this.backupPriorities = backupPriorities;
-    indexPatternsOrdered = backupPriorities.indicesSplitBySnapshot().toList();
     this.repository = repository;
     this.backupProps = backupProps;
     this.dynamicIndicesProvider = dynamicIndicesProvider;
@@ -52,13 +50,16 @@ public class BackupServiceImpl implements BackupService {
   public void deleteBackup(final Long backupId) {
     repository.validateRepositoryExists(backupProps.repositoryName());
     final String repositoryName = backupProps.repositoryName();
-    final int count = indexPatternsOrdered.size();
-    final String version = getCurrentVersion();
-    for (int index = 0; index < count; index++) {
+    // the number of parts can be dynamic
+    final var backupMetadataOpt = repository.getMetadata(repositoryName, backupId);
+    if (backupMetadataOpt.isEmpty()) {
+      throw new InvalidRequestException("Cannot find backup with id " + backupId);
+    }
+    final var backupMetadata = backupMetadataOpt.get();
+    final int partCount = backupMetadata.partCount();
+    for (int partIdx = 1; partIdx <= partCount; partIdx++) {
       final String snapshotName =
-          repository
-              .snapshotNameProvider()
-              .getSnapshotName(new Metadata(backupId, version, index + 1, count));
+          repository.snapshotNameProvider().getSnapshotName(backupMetadata.withPart(partIdx));
       repository.deleteSnapshot(repositoryName, snapshotName);
     }
   }
@@ -90,15 +91,17 @@ public class BackupServiceImpl implements BackupService {
 
   TakeBackupResponseDto scheduleSnapshots(final TakeBackupRequestDto request) {
     final String repositoryName = backupProps.repositoryName();
+    final var indexPatternsOrdered = getValidIndexPatterns();
     final int count = indexPatternsOrdered.size();
     final List<String> snapshotNames = new ArrayList<>();
     final String version = getCurrentVersion();
-    var index = 1;
-    for (var indexCollection : indexPatternsOrdered) {
-      final Metadata metadata = new Metadata(request.getBackupId(), version, index++, count);
+    for (int index = 0; index < indexPatternsOrdered.size(); index++) {
+      SnapshotIndexCollection indexCollection = indexPatternsOrdered.get(index);
+      final var partNum = index + 1;
+      final Metadata metadata = new Metadata(request.getBackupId(), version, partNum, count);
       final String snapshotName = repository.snapshotNameProvider().getSnapshotName(metadata);
       // Add all the dynamic indices in the last step
-      if (index == count + 1) {
+      if (partNum == count) {
         indexCollection =
             indexCollection.addSkippableIndices(dynamicIndicesProvider.getAllDynamicIndices());
       }
@@ -124,6 +127,28 @@ public class BackupServiceImpl implements BackupService {
                   nextRequest, false, this::scheduleNextSnapshot, requestsQueue::clear));
       LOGGER.debug("Snapshot picked for execution: {}", nextRequest);
     }
+  }
+
+  @VisibleForTesting
+  List<SnapshotIndexCollection> getValidIndexPatterns() {
+    final var list = new ArrayList<SnapshotIndexCollection>();
+    final var missingIndicesList = new ArrayList<String>();
+    for (final var indices : backupPriorities.indicesSplitBySnapshot().toList()) {
+      final var foundIndices = repository.checkAllIndicesExist(indices.allIndices());
+      final var missingRequiredIndices =
+          indices.requiredIndices().stream().filter(idx -> !foundIndices.contains(idx)).toList();
+      if (!missingRequiredIndices.isEmpty()) {
+        missingIndicesList.addAll(missingRequiredIndices);
+      }
+      // skip this part if there is no index, but they are not required
+      if (!foundIndices.isEmpty()) {
+        list.add(indices);
+      }
+    }
+    if (!missingIndicesList.isEmpty()) {
+      throw new IndexNotFoundException(missingIndicesList);
+    }
+    return list;
   }
 
   // TODO Keep in mind this bug: https://github.com/camunda/camunda/issues/20458
